+++ src/main/java/org/folio/dcb/domain/DcbAgencyKey.java
@@ -1,0 +1,3 @@
+package org.folio.dcb.domain;
+
+public record DcbAgencyKey(String agencyCode, String agencyName) {}